fix(google): Update constructors to account for retry changes (#1640)
* fix(google): Update constructors to account for retry changes

The changes to the retry logic updated some fields from Long to
int (because we won't be waiting 24 days between retries); this
commit just updates the places we called them in Halyard to pass
an int.

* chore(dependencies): Bump front50 version

+++ halyard-config/src/main/java/com/netflix/spinnaker/halyard/config/validate/v1/canary/google/GoogleCanaryAccountValidator.java
@@ -45,10 +45,10 @@ public class GoogleCanaryAccountValidator extends CanaryAccountValidator<GoogleC
 
   private int connectTimeoutSec = 45;
   private int readTimeoutSec = 45;
-  private long maxWaitInterval = 60000;
-  private long retryIntervalBase = 2;
-  private long jitterMultiplier = 1000;
-  private long maxRetries = 10;
+  private int maxWaitInterval = 60000;
+  private int retryIntervalBase = 2;
+  private int jitterMultiplier = 1000;
+  private int maxRetries = 10;
 
   GoogleCanaryAccountValidator(SecretSessionManager secretSessionManager) {
     this.secretSessionManager = secretSessionManager;
